+++ Models/SavedPosition.cs
@@ -1,0 +1,8 @@
+﻿namespace ScheduleToolbox.Models;
+
+[Serializable]
+public class SavedPosition
+{
+    public SerializableVector3 position;
+    public SerializableQuaternion rotation;
+}