+++ Models/SavedPositionData.cs
@@ -1,0 +1,7 @@
+﻿namespace ScheduleToolbox.Models;
+
+[Serializable]
+public class SavedPositionsData
+{
+    public Dictionary<string, SavedPosition> Positions = new();
+}