+++ Models/ToolboxData.cs
@@ -1,0 +1,10 @@
+﻿using UnityEngine;
+
+namespace ScheduleToolbox.Models;
+
+[Serializable]
+public class ToolboxData
+{
+    public Dictionary<string, string> Keybindings = new();
+    public Dictionary<string, SavedPosition> Positions = new();
+}